refactor/#2/Refactor ordinal date support
This change refactors the ordinal date support to
allow for more options to be used with it. For
example we now allow using ordinal dates with
time as well, e.g. "2023-095 22:18".

This was accomplished by pre-processing the date
string, converting the ordinal date from the
`yyyy-ddd` format to `yyyy-mm-dd` and then
continuing parsing as normal.